Adds post creation support.
Fixes a bug when using the option to overwrite the entire configuration
file which didn't properly validate the selected blog or update the
selected blog in the menu heading.

Adds the ability to create a new post. When creating a post, a prompt is
added to validate if the user would like to create the post, discard it,
or re-edit. When the post is completed, either due to being posted or
discarded, the local file is removed.